Added Load Game, but with incorrect stats. :(

+++ SpaceShipFarcrothu/SpaceShipFarcrothu.Models/ModelEntities/AsteroidModel.cs
@@ -9,7 +9,7 @@ namespace SpaceShipFarcrothu.Models.ModelEntities
     public class AsteroidModel
     {
         public int Id { get; set; }
-        public double PositionX { get; set; }
-        public double PositionY { get; set; }
+        public float PositionX { get; set; }
+        public float PositionY { get; set; }
     }
 }
