+++ SpaceShipFarcrothu/SpaceShipFarcrothu.Models/ModelEntities/EnemyModel.cs
@@ -3,7 +3,7 @@
     public class EnemyModel
     {
         public int Id { get; set; }
-        public double PositionX { get; set; }
-        public double PositionY { get; set; }
+        public float PositionX { get; set; }
+        public float PositionY { get; set; }
     }
 }
